Add existing Unity project with test water shader
This project is a test to combine the use of the Gerstner waves from the Water4 shader that comes with Unity, with the Standard shader. (The Water4 shader does not react to lights). At this point, the Gerstner waves are now working with the Standard shader. The ultimate goal is to allow meshes of varying shape (representing a waterfall for instance), that can accept a map to control how much and where Gerstner wave intensity is applied; that can accept a flow map for bump distortion that flows with a water channel; and that can be tesselated or used with a Level of Detail group, for smooth water with a reasonable poly count.