Add ability to accept wave and color maps, update README
The waterfall shader can now take a color map, and apply it according the the model's UV's, and take a wave map, which controls the amount of the Gerstner wave offsets that are actually applied. Also, there is a conditional which checks each vertex for whether the wave map would allow any amount of the Gerstner waves to show. If there will be no waves showing at that vertex, then the waves are not even calculated for that vertex. Besides saving on calculations, this has the additional benefit that it avoids passing on the changes in vertex position to the lighting. Otherwise, the wave shading would be showing in the flat areas where it isn't wanted.
Also, updated the README to list the goals of this project.