Modify the model to curve up, and add code to rotate the waves
The model is starting to be curved up, as in a waterfall. The waves still are displaced straight up, despite the model curving. Added code for rotating the wave offset by the amount of difference between the polygon normal, and up, using a rotation matrix. This code compiles, but has not yet been aplied to the wave shader code.