Implement wave rotation to be perpendicular to vertex normal
Hooked up the code for rotation of the offsets when the water is not flat. The magnitude of the offset vector is saved, then the offset is normalized and rotated. After the rotation, the offset is multiplied by its original saved magnitude. Added a conditional so that the rotation calculations only happen when the poly is not close to flat.